+++ source/autoCompBase.js
@@ -1,4 +1,3 @@
-// This file contains auto-completer code for the Data Entry Framework project.
 
 // These autocompleters are based on the Autocompleter.Base class defined
 // in the Script.aculo.us controls.js file.   Most of the controls.js code has
@@ -1033,11 +1032,18 @@ if (typeof Def === 'undefined')
         else
           this.index = this.entryCount-1;
         highlightedLITag = this.getCurrentEntry(); // depends on this.index
-      } while (this.liIsHeading(highlightedLITag) && this.index !== stopIndex);
+        var itemText = (this.listItemValue(highlightedLITag));
 
+        if (this.itemTextIsHeading(itemText)) {
+          Def.Autocompleter.screenReaderLog('Above list heading: '+itemText);
+          highlightedLITag = null;
+        }
+      } while (!highlightedLITag && this.index !== stopIndex);
 
-      this.scrollToShow(highlightedLITag, this.update.parentNode);
-      this.updateElementAfterMarking(highlightedLITag);
+      if (highlightedLITag) {
+        this.scrollToShow(highlightedLITag, this.update.parentNode);
+        this.updateElementAfterMarking(highlightedLITag);
+      }
     },
 
 
@@ -1062,11 +1068,19 @@ if (typeof Def === 'undefined')
         else
           this.index = 0;
         highlightedLITag = this.getCurrentEntry(); // depends on this.index
-      } while (this.liIsHeading(highlightedLITag) && this.index !== stopIndex);
+        var itemText = (this.listItemValue(highlightedLITag));
+
+        if (this.itemTextIsHeading(itemText)) {
+          Def.Autocompleter.screenReaderLog('Under list heading: '+itemText);
+          highlightedLITag = null;
+        }
+      } while (!highlightedLITag && this.index !== stopIndex);
 
 
-      this.scrollToShow(highlightedLITag, this.update.parentNode);
-      this.updateElementAfterMarking(highlightedLITag);
+      if (highlightedLITag) {
+        this.scrollToShow(highlightedLITag, this.update.parentNode);
+        this.updateElementAfterMarking(highlightedLITag);
+      }
     },
 
 
@@ -1087,6 +1101,20 @@ if (typeof Def === 'undefined')
       }
       else
         this.element.select();
+      // At least under some circumstances, JAWS reads the field value (perhaps
+      // because of the "select" above).  However, if this is a table-format
+      // autocompleter, we need to read the row.
+      if (this.options.tableFormat) {
+        var logEntry = [];
+        var cells = jQuery(listElement).children('td');
+        // Only read the row if there is more than one cell, because the screen
+        // reader will read what gets put in the field.
+        if (cells.length > 1) {
+          for (var i=0, len=cells.length; i<len; ++i)
+            logEntry.push(cells[i].innerText);
+          Def.Autocompleter.screenReaderLog(logEntry.join('; '));
+        }
+      }
     },
 
 
@@ -1120,6 +1148,10 @@ if (typeof Def === 'undefined')
       if (previouslyHidden && !this.temporaryHide_ && this.entryCount > 0) {
         Def.Autocompleter.screenReaderLog('A list has appeared below the '+
           this.getFieldName()+'.');
+        if (this.options.tableFormat && this.options.colHeaders) {
+          Def.Autocompleter.screenReaderLog('The column headers on the '+
+           'multi-column list are ' + this.options.colHeaders.join('; '));
+        }
       }
     },
 
@@ -1339,19 +1371,21 @@ if (typeof Def === 'undefined')
      *  value either matches a list item or is blank, and false otherwise.
      */
     setMatchStatusIndicator: function(matchStatus) {
-      if (matchStatus) {
-        if (jQuery(this.element).hasClass('no_match')) {
-          jQuery(this.element).removeClass('no_match');
-          Def.Autocompleter.screenReaderLog(
-            'The field no longer contains a non-matching value.');
+      if (matchStatus !== this.matchStatus_) {
+        if (matchStatus) {
+          if (jQuery(this.element).hasClass('no_match')) {
+            jQuery(this.element).removeClass('no_match');
+            Def.Autocompleter.screenReaderLog(
+              'The field no longer contains a non-matching value.');
+          }
         }
+        else {
+          jQuery(this.element).addClass('no_match');
+          Def.Autocompleter.screenReaderLog(
+            'The field\'s value does not match any items in the list.');
+        }
+        this.matchStatus_ = matchStatus;
       }
-      else {
-        jQuery(this.element).addClass('no_match');
-        Def.Autocompleter.screenReaderLog(
-          'The field\'s value does not match any items in the list.');
-      }
-      this.matchStatus_ = matchStatus;
     },
 
 
@@ -1487,7 +1521,7 @@ if (typeof Def === 'undefined')
         var innerMatchMinLength = minLength;
 
         for (var i=0; i<numItems; ++i) {
-          // Make sure the entry is not a header before considering it
+          // Make sure the entry is not a heading before considering it
           var itemText = listItems[i];
           if (!this.itemTextIsHeading(itemText)) {
             var itemTextLC = itemText.toLowerCase();
@@ -1938,8 +1972,13 @@ if (typeof Def === 'undefined')
     /**
      *  Runs the stuff that needs to be run when the field changes.  (This assumes
      *  that the field has changed.)
+     * @param matchStatus (optional) Set this to false if this should assume the
+     *  field value does not match the list.  If not provided, this.matchStatus_
+     *  will be used.
      */
-    propagateFieldChanges: function() {
+    propagateFieldChanges: function(matchStatus) {
+      if (matchStatus === undefined)
+        matchStatus = this.matchStatus_;
       // If this autocompleter has a record data requester, run it or clear
       // the output fields.  This will make sure the output fields are clear
       // before the change event observers run for this field, in case one of
@@ -1947,7 +1986,7 @@ if (typeof Def === 'undefined')
       // fields.  (If it does, it can wait for the record data requester's
       // latestPendingAjaxRequest_ variable to be null.)
       if (this.recDataRequester_) {
-        if (this.matchStatus_ && this.domCache.get('elemVal').trim() !== '')
+        if (matchStatus && this.domCache.get('elemVal').trim() !== '')
           this.recDataRequester_.requestData();
         else // no data, or no data from list
           this.recDataRequester_.clearDataOutputFields();
@@ -2074,10 +2113,7 @@ if (typeof Def === 'undefined')
      *  that is not a list item.
      */
     handleNonListEntry: function() {
-      // Set the match status to false, so the propagateFieldChanges will do
-      // the right thing.
-      this.matchStatus_ = false;
-      this.propagateFieldChanges();
+      this.propagateFieldChanges(false);
 
       // For a single selection list, clear the stored selection
       if (!this.multiSelect_) {