+++ source/autoCompPrefetch.js
@@ -320,6 +320,8 @@
       var lastHeading = null;
       var foundItemForLastHeading = false;
       var headerCount = 0;
+      var headingsShown = 0;
+      var skippedSelected = 0; // items already selected that are left out of the list
       var escapeHTML = Def.Autocompleter.Base.escapeAttribute;
       if (instance.options.ignoreCase)
         entry = entry.toLowerCase();
@@ -346,14 +348,16 @@
           // list numbers.
           // See if the entry matches a number.
           var itemNumStr = null;
-          var matchesItemNum = false;
+          var matchesItemNum = false; // exact match
+          var matchInItemNum = false; // partial match
           if (instance.add_seqnum) {
             itemNumStr = (i+1-headerCount)+'';
             var isSelectedByNumber = (itemNumStr === entry);
             if (!useFullList &&
                 (isSelectedByNumber || itemNumStr.indexOf(entry) === 0)) {
               ++totalCount;
-              if (isSelectedByNumber || totalCount < maxReturn) {
+              matchInItemNum = true;
+              if (isSelectedByNumber || totalCount <= maxReturn) {
                 itemNumStr = '<strong>' + itemNumStr.substr(0, entry.length) +
                   '</strong>' + itemNumStr.substr(entry.length);
                 matchesItemNum = true;
@@ -363,7 +367,7 @@
             }
           } // if we're adding sequence numbers to this list
 
-          if (!matchesItemNum && !useFullList) {
+          if (!matchInItemNum && !useFullList) {
             // See if it matches the item at the beginning
             var foundMatch = false;
             var elemComp = rawItemText;
@@ -371,44 +375,53 @@
               elemComp = rawItemText.toLowerCase();
             var foundPos = elemComp.indexOf(entry);
             while (!foundMatch && foundPos !== -1) {
-              if (totalCount < maxReturn) {
-                if (foundPos === 0) {
-                  ++totalCount;
+              if (foundPos === 0) {
+                ++totalCount;
+                foundMatch = true;
+                if (totalCount <= maxReturn) {
                   itemText = '<strong>' +
                     escapeHTML(rawItemText.substr(0, entry.length))+'</strong>'+
                     escapeHTML(rawItemText.substr(entry.length));
-                  foundMatch = true;
                 }
-                else { // foundPos > 0
-                  // See if the match is at a word boundary
-                  if (instance.options.fullSearch ||
+              }
+              else { // foundPos > 0
+                // See if the match is at a word boundary
+                if (instance.options.fullSearch ||
                     /(.\b|_)./.test(elemComp.substr(foundPos-1,2))) {
-                    ++totalCount;
+                  ++totalCount;
+                  foundMatch = true;
+                  if (totalCount <= maxReturn) {
                     var prefix = escapeHTML(rawItemText.substr(0, foundPos));
                     itemText = prefix + '<strong>' +
                       escapeHTML(rawItemText.substr(foundPos, entry.length)) +
                      '</strong>' +
                       escapeHTML(rawItemText.substr(foundPos + entry.length));
-                    foundMatch = true;
                   }
                 }
-              } // if we haven't exceeded maxReturn
+              }
               if (!foundMatch)
                 foundPos =  elemComp.indexOf(entry, foundPos+1);
             } // while we haven't found a match at a word boundary
           } // if it didn't match the item number
 
+          var alreadySelected = false;
+          if (instance.multiSelect_) {
+            alreadySelected = instance.isSelected(rawItemText)
+            if (alreadySelected)
+              ++skippedSelected;
+          }
           // Make sure that if the item's number is an exact match for what was
           // typed, it gets into the list (unless already selected).
 
           // For multi-select lists, filter out currently selected items.
           // Then, only add it if we haven't exceeded the limit.
-          if ((!instance.multiSelect_ || !instance.isSelected(rawItemText)) &&
-              itemText && (isSelectedByNumber || totalCount <= maxReturn ||
+          if (!alreadySelected && itemText &&
+              (isSelectedByNumber || totalCount <= maxReturn ||
                             (instance.numHeadings_>0 && useFullList))) {
             if (lastHeading && !foundItemForLastHeading) {
               foundItemForLastHeading = true;
               itemsInList.push(lastHeading);
+              ++headingsShown;
               itemToHTMLData[lastHeading] = [escapeHTML(lastHeading), 'heading'];
               countForLastHeading = 0;
             }
@@ -429,8 +442,9 @@
         } // else this is not a heading
       } // for each item
 
-      if (totalCount > maxReturn) {
-        $('searchCount').innerHTML = maxReturn + ' of ' + totalCount +
+      var itemsShownCount = itemsInList.length - headingsShown;
+      if (totalCount > itemsShownCount + skippedSelected) {
+        $('searchCount').innerHTML = itemsShownCount + ' of ' + totalCount +
           ' items total';
         $('moreResults').style.display = 'block';
         $('searchCount').style.display = 'block';